Prepare for migration

+++ src/elements/common-elements/file-attachments-tab/file-attachments-tab.js
@@ -139,8 +139,6 @@
         },
 
         listeners: {
-            'dialog-confirmed': '_sendRequest',
-            'delete-confirmed': '_sendRequest',
             'attachments-request-completed': '_requestCompleted'
         },
 
@@ -271,19 +269,13 @@
             });
         },
 
-        _sendRequest: function(e) {
-            if (!this.dialogOpened || !this.validate()) { return; }
+        _saveAttachment: function(e) {
+            if (!this.validate()) { return; }
 
             this.requestInProcess = true;
-            let attachmentsData, method;
 
-            if (this.deleteDialog) {
-                attachmentsData = {id: this.editedItem.id};
-                method = 'DELETE';
-            } else {
-                attachmentsData = this._getFileData();
-                method = attachmentsData.id ? 'PATCH' : 'POST';
-            }
+            let attachmentsData = this._getFileData();
+            let method = attachmentsData.id ? 'PATCH' : 'POST';
 
             attachmentsData = method === 'PATCH' ? this._getChanges(attachmentsData) : attachmentsData;
             if (!attachmentsData) {
@@ -293,6 +285,21 @@
             this.requestData = {method, attachmentsData};
         },
 
+        _deleteAttachment(event) {
+            if (this.deleteCanceled(event)) {
+                return;
+            }
+            if (!this.baseId) {
+                this._processDelayedRequest();
+                return;
+            }
+            this.requestInProcess = true;
+
+            this.requestData = {
+                method: 'DELETE',
+                attachmentsData: {id: this.editedItem.id}
+            };
+        },
         _getChanges: function(attachmentsData) {
             let original = this.originalEditedObj && this._getFileData(false, this.originalEditedObj);
             if (!original) { return attachmentsData; }
@@ -515,9 +522,13 @@
             this.deleteLinkOpened = true;
         },
 
-        _removeLink: function ({ detail }) {
+        _removeLink: function (event) {
+            if (this.deleteCanceled(event)) {
+                return;
+            }
+
             this.deleteLinkOpened = false;
-            const id = detail.dialogName;
+            const id = Polymer.dom(event).localTarget.getAttribute('link-id');
 
             this.sendRequest({
                 method: 'DELETE',
